requirements.txt
termcolor
jira-python 
suds
